requirements.txt
selenium==3.141.0
PyYAML==5.3.1
pandas==1.1.3